Initial checkin of score converter
This app replaces the current iOS Tichu Guru application. We extracted the PList from an iPhone unencrypted backup to save our current scores. This little tool converts to a JSON file that will be imported by the replacement app we write.

This app is definitely overkill for the purpose; I mainly wanted to play around with the PList format.