Add iOS Tichu App
This is a rough prototype so far. Currently one can create a new game with some hard-coded players and view the newly created game on the game list (but cannot go into the game's hand scores yet). The main things so far are a navigation hierarchy and the ability to pop up new windows and return a result to the parent window.

The data model is basically complete and just needs serialization adding.